fix: second wave of SonarCloud rule fixes across Java/Kotlin/XML

+++ app/src/main/java/com/simplecity/amp_library/playback/PlaybackManager.java
@@ -340,7 +340,7 @@ public class PlaybackManager implements Playback.Callbacks {
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(songs -> {
                     if (!songs.isEmpty() && queueManager.getCurrentSong() != null) {
-                        load(queueManager.getCurrentSong(), playWhenReady, (long) 0, null);
+                        load(queueManager.getCurrentSong(), playWhenReady, 0L, null);
                     }
                 }, error -> LogUtils.logException(TAG, "Error opening file", error)));
     }
